Support per operation sampling (#112)
Add a PerOperationSampler that maintains a sampling probability for every operation

+++ jaeger-core/src/main/java/com/uber/jaeger/samplers/RateLimitingSampler.java
@@ -35,10 +35,10 @@ public class RateLimitingSampler implements Sampler {
   public static final String TYPE = "ratelimiting";
 
   private final RateLimiter rateLimiter;
-  private final int maxTracesPerSecond;
+  private final double maxTracesPerSecond;
   private final Map<String, Object> tags;
 
-  public RateLimitingSampler(int maxTracesPerSecond) {
+  public RateLimitingSampler(double maxTracesPerSecond) {
     this.maxTracesPerSecond = maxTracesPerSecond;
     this.rateLimiter = new RateLimiter(maxTracesPerSecond);
